+++ packages/store-sync/src/SyncStep.ts
@@ -1,0 +1,6 @@
+export enum SyncStep {
+  INITIALIZE = "initialize",
+  SNAPSHOT = "snapshot",
+  RPC = "rpc",
+  LIVE = "live",
+}
