+++ packages/store-sync/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./blockLogsToStorage";
 export * from "./common";
+export * from "./SyncStep";
