+++ packages/store-sync/src/recs/common.ts
@@ -4,10 +4,3 @@ export type StoreComponentMetadata = {
   keySchema: KeySchema;
   valueSchema: ValueSchema;
 };
-
-export enum SyncStep {
-  INITIALIZE = "initialize",
-  SNAPSHOT = "snapshot",
-  RPC = "rpc",
-  LIVE = "live",
-}
