+++ packages/store-sync/src/recs/syncToRecs.ts
@@ -1,5 +1,4 @@
-import { StoreConfig, storeEventsAbi } from "@latticexyz/store";
-import { Hex, TransactionReceipt } from "viem";
+import { StoreConfig } from "@latticexyz/store";
 import {
   ComponentValue,
   Entity,
@@ -10,22 +9,16 @@ import {
   setComponent,
 } from "@latticexyz/recs";
 import { SyncOptions, SyncResult } from "../common";
-import {
-  createBlockStream,
-  isNonPendingBlock,
-  blockRangeToLogs,
-  groupLogsByBlockNumber,
-} from "@latticexyz/block-logs-stream";
-import { filter, map, tap, mergeMap, from, concatMap, share, firstValueFrom } from "rxjs";
-import { blockLogsToStorage } from "../blockLogsToStorage";
 import { recsStorage } from "./recsStorage";
 import { hexKeyTupleToEntity } from "./hexKeyTupleToEntity";
 import { debug } from "./debug";
 import { defineInternalComponents } from "./defineInternalComponents";
 import { getTableKey } from "./getTableKey";
-import { StoreComponentMetadata, SyncStep } from "./common";
+import { StoreComponentMetadata } from "./common";
+import { SyncStep } from "../SyncStep";
 import { encodeEntity } from "./encodeEntity";
 import { createIndexerClient } from "../trpc-indexer";
+import { createStoreSync } from "../createStoreSync";
 
 type SyncToRecsOptions<
   TConfig extends StoreConfig = StoreConfig,
@@ -133,96 +126,56 @@ export async function syncToRecs<
     });
   }
 
-  // TODO: if startBlock is still 0, find via deploy event
-
-  debug("starting sync from block", startBlock);
-
-  const latestBlock$ = createBlockStream({ publicClient, blockTag: "latest" }).pipe(share());
-
-  const latestBlockNumber$ = latestBlock$.pipe(
-    filter(isNonPendingBlock),
-    map((block) => block.number),
-    share()
-  );
+  const storeSync = await createStoreSync({
+    storageAdapter: recsStorage({ components, config }),
+    config,
+    address,
+    publicClient,
+    startBlock,
+    maxBlockRange,
+    indexerUrl,
+    initialState,
+  });
 
   let latestBlockNumber: bigint | null = null;
-  const blockLogs$ = latestBlockNumber$.pipe(
-    tap((blockNumber) => {
-      debug("latest block number", blockNumber);
+  {
+    const sub = storeSync.latestBlockNumber$.subscribe((blockNumber) => {
       latestBlockNumber = blockNumber;
-    }),
-    map((blockNumber) => ({ startBlock, endBlock: blockNumber })),
-    blockRangeToLogs({
-      publicClient,
-      address,
-      events: storeEventsAbi,
-      maxBlockRange,
-    }),
-    mergeMap(({ toBlock, logs }) => from(groupLogsByBlockNumber(logs, toBlock))),
-    share()
-  );
-
-  let lastBlockNumberProcessed: bigint | null = null;
-  const blockStorageOperations$ = blockLogs$.pipe(
-    concatMap(blockLogsToStorage(recsStorage({ components, config }))),
-    tap(({ blockNumber, operations }) => {
-      debug("stored", operations.length, "operations for block", blockNumber);
-      lastBlockNumberProcessed = blockNumber;
-
-      if (
-        latestBlockNumber != null &&
-        getComponentValue(components.SyncProgress, singletonEntity)?.step !== SyncStep.LIVE
-      ) {
-        if (blockNumber < latestBlockNumber) {
-          setComponent(components.SyncProgress, singletonEntity, {
-            step: SyncStep.RPC,
-            message: `Hydrating from RPC to block ${latestBlockNumber}`,
-            percentage: (Number(blockNumber) / Number(latestBlockNumber)) * 100,
-          });
-        } else {
-          setComponent(components.SyncProgress, singletonEntity, {
-            step: SyncStep.LIVE,
-            message: `All caught up!`,
-            percentage: 100,
-          });
-        }
-      }
-    }),
-    share()
-  );
-
-  // Start the sync
-  const sub = blockStorageOperations$.subscribe();
-  world.registerDisposer(() => sub.unsubscribe());
-
-  async function waitForTransaction(tx: Hex): Promise<{
-    receipt: TransactionReceipt;
-  }> {
-    // Wait for tx to be mined
-    const receipt = await publicClient.waitForTransactionReceipt({ hash: tx });
-
-    // If we haven't processed a block yet or we haven't processed the block for the tx, wait for it
-    if (lastBlockNumberProcessed == null || lastBlockNumberProcessed < receipt.blockNumber) {
-      await firstValueFrom(
-        blockStorageOperations$.pipe(
-          filter(({ blockNumber }) => blockNumber != null && blockNumber >= receipt.blockNumber)
-        )
-      );
-    }
-
-    return { receipt };
+    });
+    world.registerDisposer(sub.unsubscribe);
   }
 
+  // Start the sync
+  let lastBlockNumberProcessed: bigint | null = null;
+  const sub = storeSync.blockStorageOperations$.subscribe(({ blockNumber, operations }) => {
+    debug("stored", operations.length, "operations for block", blockNumber);
+    lastBlockNumberProcessed = blockNumber;
+
+    if (
+      latestBlockNumber != null &&
+      getComponentValue(components.SyncProgress, singletonEntity)?.step !== SyncStep.LIVE
+    ) {
+      if (blockNumber < latestBlockNumber) {
+        setComponent(components.SyncProgress, singletonEntity, {
+          step: SyncStep.RPC,
+          message: `Hydrating from RPC to block ${latestBlockNumber}`,
+          percentage: (Number(blockNumber) / Number(latestBlockNumber)) * 100,
+        });
+      } else {
+        setComponent(components.SyncProgress, singletonEntity, {
+          step: SyncStep.LIVE,
+          message: `All caught up!`,
+          percentage: 100,
+        });
+      }
+    }
+  });
+
+  world.registerDisposer(sub.unsubscribe);
+
   return {
+    ...storeSync,
     components,
     singletonEntity,
-    latestBlock$,
-    latestBlockNumber$,
-    blockLogs$,
-    blockStorageOperations$,
-    waitForTransaction,
-    destroy: (): void => {
-      world.dispose();
-    },
   };
 }
