+++ packages/store-sync/src/sqlite/syncToSqlite.ts
@@ -1,17 +1,8 @@
-import { Hex, TransactionReceipt } from "viem";
-import {
-  createBlockStream,
-  isNonPendingBlock,
-  blockRangeToLogs,
-  groupLogsByBlockNumber,
-} from "@latticexyz/block-logs-stream";
-import { concatMap, filter, firstValueFrom, from, map, mergeMap, share, tap } from "rxjs";
-import { StoreConfig, storeEventsAbi } from "@latticexyz/store";
+import { StoreConfig } from "@latticexyz/store";
 import { BaseSQLiteDatabase } from "drizzle-orm/sqlite-core";
-import { debug } from "../debug";
 import { SyncOptions, SyncResult } from "../common";
-import { blockLogsToStorage } from "../blockLogsToStorage";
 import { sqliteStorage } from "./sqliteStorage";
+import { createStoreSync } from "../createStoreSync";
 
 type SyncToSqliteOptions<TConfig extends StoreConfig = StoreConfig> = SyncOptions<TConfig> & {
   /**
@@ -22,7 +13,9 @@ type SyncToSqliteOptions<TConfig extends StoreConfig = StoreConfig> = SyncOption
   database: BaseSQLiteDatabase<"sync", any>;
 };
 
-type SyncToSqliteResult<TConfig extends StoreConfig = StoreConfig> = SyncResult<TConfig>;
+type SyncToSqliteResult<TConfig extends StoreConfig = StoreConfig> = SyncResult<TConfig> & {
+  destroy: () => void;
+};
 
 /**
  * Creates an indexer to process and store blockchain events.
@@ -30,7 +23,8 @@ type SyncToSqliteResult<TConfig extends StoreConfig = StoreConfig> = SyncResult<
  * @param {CreateIndexerOptions} options See `CreateIndexerOptions`.
  * @returns A function to unsubscribe from the block stream, effectively stopping the indexer.
  */
-export function syncToSqlite<TConfig extends StoreConfig = StoreConfig>({
+export async function syncToSqlite<TConfig extends StoreConfig = StoreConfig>({
+  config,
   database,
   publicClient,
   address,
@@ -38,76 +32,23 @@ export function syncToSqlite<TConfig extends StoreConfig = StoreConfig>({
   maxBlockRange,
   indexerUrl,
   initialState,
-}: SyncToSqliteOptions<TConfig>): SyncToSqliteResult<TConfig> {
-  // TODO: sync initial state
-
-  const latestBlock$ = createBlockStream({ publicClient, blockTag: "latest" }).pipe(share());
-
-  const latestBlockNumber$ = latestBlock$.pipe(
-    filter(isNonPendingBlock),
-    map((block) => block.number),
-    share()
-  );
-
-  let latestBlockNumber: bigint | null = null;
-  const blockLogs$ = latestBlockNumber$.pipe(
-    tap((blockNumber) => {
-      debug("latest block number", blockNumber);
-      latestBlockNumber = blockNumber;
-    }),
-    map((blockNumber) => ({ startBlock, endBlock: blockNumber })),
-    blockRangeToLogs({
-      publicClient,
-      address,
-      events: storeEventsAbi,
-      maxBlockRange,
-    }),
-    tap(({ fromBlock, toBlock, logs }) => {
-      debug("found", logs.length, "logs for block", fromBlock, "-", toBlock);
-    }),
-    mergeMap(({ toBlock, logs }) => from(groupLogsByBlockNumber(logs, toBlock))),
-    share()
-  );
-
-  let lastBlockNumberProcessed: bigint | null = null;
-  const blockStorageOperations$ = blockLogs$.pipe(
-    concatMap(blockLogsToStorage(sqliteStorage({ database, publicClient }))),
-    tap(({ blockNumber, operations }) => {
-      debug("stored", operations.length, "operations for block", blockNumber);
-      lastBlockNumberProcessed = blockNumber;
-
-      // TODO: store some notion of sync progress?
-    }),
-    share()
-  );
+}: SyncToSqliteOptions<TConfig>): Promise<SyncToSqliteResult<TConfig>> {
+  const storeSync = await createStoreSync({
+    storageAdapter: sqliteStorage({ database, publicClient }),
+    config,
+    address,
+    publicClient,
+    startBlock,
+    maxBlockRange,
+    indexerUrl,
+    initialState,
+  });
 
   // Start the sync
-  const sub = blockStorageOperations$.subscribe();
-
-  async function waitForTransaction(tx: Hex): Promise<{
-    receipt: TransactionReceipt;
-  }> {
-    // Wait for tx to be mined
-    const receipt = await publicClient.waitForTransactionReceipt({ hash: tx });
-
-    // If we haven't processed a block yet or we haven't processed the block for the tx, wait for it
-    if (lastBlockNumberProcessed == null || lastBlockNumberProcessed < receipt.blockNumber) {
-      await firstValueFrom(
-        blockStorageOperations$.pipe(
-          filter(({ blockNumber }) => blockNumber != null && blockNumber >= receipt.blockNumber)
-        )
-      );
-    }
-
-    return { receipt };
-  }
+  const sub = storeSync.blockStorageOperations$.subscribe();
 
   return {
-    latestBlock$,
-    latestBlockNumber$,
-    blockLogs$,
-    blockStorageOperations$,
-    waitForTransaction,
+    ...storeSync,
     destroy: (): void => {
       sub.unsubscribe();
     },
